requirements.txt
# Core
sentence-transformers>=3.0.0
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0
umap-learn>=0.5.6
faiss-cpu>=1.8.0.post1
tqdm>=4.66.0
# Bonus app
streamlit>=1.34.0
transformers>=4.41.0
PyPDF2>=3.0.0
pypdf>=4.2.0
